// src/types/salesOrderData.ts
export default interface SalesOrderData {
      Vbeln: string;
      Posnr: string;
      Matnr: string;
      Arktx: string;
      Auart: string;
      Erdat: string;
      Kwmeng: string;
      VdatuAna: string;
      Bstnk: string;
      Kunnr: string;
      Vrkme: string;
      Netwr: string;
      Waerk: string;
      Spart: string;
      Gbstk: string;
      Lfgsk: string;
      Lgort: string;
};